Rework FFINDEX_STATIC.
Libffindex.a is now always built and installed.
Static programs are conditionalized on FFINDEX_STATIC,
with the exception of MPI programms and Mac OS X,
where this might be discouraged[1][2].

For MPI one could add a FFINDEX_MPI_STATIC flag.
Building crt0.o for Mac OS X would be left to the user.

[1] http://www.open-mpi.org/faq/?category=mpi-apps#static-mpi-apps
[2] https://developer.apple.com/library/mac/qa/qa1118/_index.html